pyproject.toml
[tool.poetry]
name = "nonebot-plugin-blacklist"
version = "1.3.1"
description = "Blacklist in NoneBot2"
authors = ["月ヶ瀬"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "nonebot_plugin_blacklist" },
]
homepage = "https://github.com/tkgs0/nonebot-plugin-blacklist"
repository = "https://github.com/tkgs0/nonebot-plugin-blacklist"
keywords = ["nonebot", "blacklist"]

[tool.poetry.dependencies]
python = "^3.8"
nonebot-adapter-onebot = "^2.2.0"
nonebot2 = "^2.0.1"
ujson = ">=5.2.0"
nonebot-plugin-orm = ">=0.5.0, <1.0.0"
redis = ">= 5.0.0"

[tool.poetry.dev-dependencies]
nonebot-plugin-orm = {extras = ["default"], version = "^0.5.0"}

[tool.mypy]
python_version = "3.8"
ignore_missing_imports = true
implicit_reexport = true
pretty = true
show_error_codes = true
strict = true

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

